requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
numpy==2.1.3
